Fixed Merge errors
