+++ frontend/huishoudappfrontend/lib/home_page.dart
@@ -1,15 +1,34 @@
 import 'package:flutter/material.dart';
+import 'package:huishoudappfrontend/createaccount_page.dart';
+import 'package:huishoudappfrontend/setup/provider.dart';
+import 'package:huishoudappfrontend/setup/auth.dart';
+import 'package:huishoudappfrontend/setup/validators.dart';
+
 
 class HomePage extends StatelessWidget {
   static String tag = 'home-page';
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Center(
-        child: Text(
-          'Home-page',
-          style: TextStyle(fontSize: 25),
+      appBar: AppBar(
+        title: Text('Welcome Page'),
+        actions: <Widget>[
+          FlatButton(
+            child: Text("Sign Out"),
+            onPressed: () async {
+              try {
+                Auth auth = Provider.of(context).auth;
+                await auth.signOut();
+              } catch (e) {
+                print(e);
+              }
+            },
+          )
+        ],
+      ),
+      body: Container(
+        child: Center(
+          child: Text('Welcome'),
         ),
       ),
     );
