+++ frontend/huishoudappfrontend/lib/login_page.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_signin_button/button_list.dart';
-import 'package:flutter_signin_button/flutter_signin_button.dart';
 import 'package:huishoudappfrontend/createaccount_page.dart';
+import 'package:huishoudappfrontend/setup/provider.dart';
+import 'package:huishoudappfrontend/setup/auth.dart';
+import 'package:huishoudappfrontend/setup/validators.dart';
 
-import 'home_page.dart';
-import 'createaccount_page.dart';
 
 class LoginPage extends StatefulWidget {
   static String tag = 'login-page';
@@ -13,92 +12,145 @@ class LoginPage extends StatefulWidget {
 }
 
 class _LoginPageState extends State<LoginPage> {
-  var x = 0;
-  increase() {
-    setState(() {
-      x++;
-    });
+  final formKey = GlobalKey<FormState>();
+
+  String _email, _password;
+  FormType _formType = FormType.login;
+
+  bool validate() {
+    final form = formKey.currentState;
+    form.save();
+    if (form.validate()) {
+      form.save();
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  void submit() async {
+    if (validate()) {
+      try {
+        final auth = Provider.of(context).auth;
+        if (_formType == FormType.login) {
+          String userId = await auth.signInWithEmailAndPassword(
+            _email,
+            _password,
+          );
+
+          print('Signed in $userId');
+        } else {
+          String userId = await auth.createUserWithEmailAndPassword(
+            _email,
+            _password,
+          );
+
+          print('Registered in $userId');
+        }
+      } catch (e) {
+        print(e);
+      }
+    }
+  }
+
+  void switchFormState(String state) {
+    formKey.currentState.reset();
+
+    if (state == 'register') {
+      setState(() {
+        _formType = FormType.register;
+      });
+    } else {
+      setState(() {
+        _formType = FormType.login;
+      });
+    }
+  }
+
+  List<Widget> buildInputs() {
+    return [
+      TextFormField(
+        validator: EmailValidator.validate,
+        decoration: InputDecoration(labelText: 'Email'),
+        onSaved: (value) => _email = value,
+      ),
+      TextFormField(
+        validator: PasswordValidator.validate,
+        decoration: InputDecoration(labelText: 'Password'),
+        obscureText: true,
+        onSaved: (value) => _password = value,
+      ),
+    ];
+  }
+
+  List<Widget> buildButtons() {
+    if (_formType == FormType.login) {
+      return [
+        RaisedButton(
+          child: Text('Login'),
+          color: Colors.blueAccent,
+          onPressed: submit,
+        ),
+        FlatButton(
+          child: Text('Register Account'),
+          color: Colors.teal,
+          onPressed: () {
+            switchFormState('register');
+          },
+        ),
+        Divider(
+          height: 50.0,
+        ),
+        FlatButton(
+          child: Text("Sign in with Google"),
+          color: Colors.lightGreen,
+          onPressed: () async {
+            try {
+              final _auth = Provider.of(context).auth;
+              final id = await _auth.signInWithGoogle();
+              print('signed in with google $id');
+            } catch (e) {
+              print(e);
+            }
+          },
+        ),
+      ];
+    } else {
+      return [
+        RaisedButton(
+          child: Text('Create Account'),
+          color: Colors.blueAccent,
+          onPressed: submit,
+        ),
+        FlatButton(
+          child: Text('Go to Login'),
+          color: Colors.teal,
+          onPressed: () {
+            switchFormState('login');
+          },
+        )
+      ];
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    final logo = Image.asset(
-      'images/testlogoapp.png',
-      width: 150,
-      height: 180,
-    );
-
-    final email = TextFormField(
-      keyboardType: TextInputType.emailAddress,
-      decoration: InputDecoration(
-        hintText: 'Gebruikersnaam',
-        contentPadding: EdgeInsets.fromLTRB(20.0, 10.0, 20.0, 10.0),
-        border: OutlineInputBorder(borderRadius: BorderRadius.circular(32.0)),
-      ),
-    );
-
-    final password = TextFormField(
-      keyboardType: TextInputType.visiblePassword,
-      decoration: InputDecoration(
-        hintText: 'Wachtwoord',
-        contentPadding: EdgeInsets.fromLTRB(20.0, 10.0, 20.0, 10.0),
-        border: OutlineInputBorder(borderRadius: BorderRadius.circular(32.0)),
-      ),
-    );
-
-    final loginButton = Padding(
-      padding: EdgeInsets.symmetric(vertical: 16.0),
-      child: RaisedButton(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
-        onPressed: () {
-          Navigator.of(context).pushNamed(HomePage.tag);
-        },
-        padding: EdgeInsets.all(12),
-        color: Colors.lightBlueAccent,
-        child: Text('Log In', style: TextStyle(color: Colors.white)),
-      ),
-    );
-
-    final googleLogIn = SignInButton(
-      Buttons.Google,
-      onPressed: increase,
-    );
-
-    final signIn = FlatButton(
-      onPressed: () {
-        Navigator.of(context).pushNamed(CreateAccount.tag);
-      },
-      child: Text(
-        'Account aanmaken',
-        textAlign: TextAlign.center,
-        style: TextStyle(
-          fontWeight: FontWeight.bold,
-          fontSize: 17,
-        ),
-      ),
-    );
-
     return Scaffold(
+      appBar: AppBar(
+        title: Text('Form Page'),
+      ),
+      resizeToAvoidBottomPadding: false,
       body: Center(
-        child: ListView(
-          padding: EdgeInsets.only(left: 40.0, right: 40.0),
-          children: <Widget>[
-            SizedBox(height: 70),
-            logo,
-            SizedBox(height: 50),
-            email,
-            SizedBox(height: 10),
-            password,
-            SizedBox(height: 10),
-            loginButton,
-            googleLogIn,
-            SizedBox(height: 8),
-            SizedBox(height: 40),
-            signIn,
-          ],
+        child: Form(
+          key: formKey,
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            mainAxisAlignment: MainAxisAlignment.center,
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: buildInputs() + buildButtons(),
+          ),
         ),
       ),
-      backgroundColor: Colors.blue[50],
     );
   }
 }
